isGrounded fix
Not sure exactly what was going on with my isGrounded check, but it wasn't rayCasting properly - the tutorial I had followed showed a setup for a 'boxCast' which I used from the foot collider, but it was too clunky, hitting off the player itself and the opponent. Changed to a proper raycast, made sure the ray was hitting the stage collider (by name, so it didn't hit the bodies of players) and adjusting the isGrounded checks!  TL/DR: The player will no longer hover in falling when actually grounded! :D